requirements.txt
aws-cdk-lib==2.138.0
constructs>=10.0.0,<11.0.0
weasyprint==59.0
Jinja2==3.1.2